chore: lint all code and fix style issues
