+++ packages/avatar/lib/AvatarImage.tsx
@@ -1,17 +1,18 @@
-import {Norn} from "@norns-ui/norn";
-import {forwardRef} from "react";
-import {ScopedProps, useAvatarContext} from "./Avatar";
 import {
   ImageLoadingStatus,
   useCallbackRef,
   useImageLoadingStatus,
   useLayoutEffect,
 } from "@norns-ui/hooks";
+import {Norn} from "@norns-ui/norn";
+import {ComponentPropsWithoutRef, ElementRef, forwardRef} from "react";
+
+import {ScopedProps, useAvatarContext} from "./Avatar";
 
 const IMAGE_NAME = "AvatarImage";
 
-type AvatarImageElement = React.ElementRef<typeof Norn.img>;
-type NornImageProps = React.ComponentPropsWithoutRef<typeof Norn.img>;
+type AvatarImageElement = ElementRef<typeof Norn.img>;
+type NornImageProps = ComponentPropsWithoutRef<typeof Norn.img>;
 interface AvatarImageProps extends NornImageProps {
   onLoadingStatusChange?: (status: ImageLoadingStatus) => void;
 }
