+++ packages/hooks/lib/useCallbackRef/useCallbackRef.ts
@@ -1,6 +1,6 @@
 "use client";
 
-import {useRef, useEffect, useMemo} from "react";
+import {useEffect, useMemo, useRef} from "react";
 
 /**
  * A custom hook that converts a callback to a ref to avoid triggering re-renders when passed as a
