+++ packages/navigation-menu/lib/index.ts
@@ -1,26 +1,26 @@
 "use client";
 
-export {FocusGroup} from "./FocusGroup";
 export type {FocusGroupProps} from "./FocusGroup";
-export {FocusGroupItem} from "./FocusGroupItem";
+export {FocusGroup} from "./FocusGroup";
 export type {FocusGroupItemProps} from "./FocusGroupItem";
-export {NavigationMenu} from "./NavigationMenu";
+export {FocusGroupItem} from "./FocusGroupItem";
 export type {NavigationMenuProps} from "./NavigationMenu";
-export {NavigationMenuContent} from "./NavigationMenuContent";
+export {NavigationMenu} from "./NavigationMenu";
 export type {NavigationMenuContentProps} from "./NavigationMenuContent";
-export {NavigationMenuIndicator} from "./NavigationMenuIndicator";
+export {NavigationMenuContent} from "./NavigationMenuContent";
 export type {NavigationMenuIndicatorProps} from "./NavigationMenuIndicator";
-export {NavigationMenuItem} from "./NavigationMenuItem";
+export {NavigationMenuIndicator} from "./NavigationMenuIndicator";
 export type {NavigationMenuItemProps} from "./NavigationMenuItem";
-export {NavigationMenuLink} from "./NavigationMenuLink";
+export {NavigationMenuItem} from "./NavigationMenuItem";
 export type {NavigationMenuLinkProps} from "./NavigationMenuLink";
-export {NavigationMenuList} from "./NavigationMenuList";
+export {NavigationMenuLink} from "./NavigationMenuLink";
 export type {NavigationMenuListProps} from "./NavigationMenuList";
-export {NavigationMenuProvider} from "./NavigationMenuProvider";
+export {NavigationMenuList} from "./NavigationMenuList";
 export type {NavigationMenuProviderProps} from "./NavigationMenuProvider";
-export {NavigationMenuSub} from "./NavigationMenuSub";
+export {NavigationMenuProvider} from "./NavigationMenuProvider";
 export type {NavigationMenuSubProps} from "./NavigationMenuSub";
-export {NavigationMenuTrigger} from "./NavigationMenuTrigger";
+export {NavigationMenuSub} from "./NavigationMenuSub";
 export type {NavigationMenuTriggerProps} from "./NavigationMenuTrigger";
-export {NavigationMenuViewport} from "./NavigationMenuViewport";
+export {NavigationMenuTrigger} from "./NavigationMenuTrigger";
 export type {NavigationMenuViewportProps} from "./NavigationMenuViewport";
+export {NavigationMenuViewport} from "./NavigationMenuViewport";
