+++ packages/presence/lib/getElementRef.ts
@@ -1,22 +1,18 @@
 import {ReactElement} from "react";
 
-// Access the ref using the method that doesn't yield a warning.
 const getElementRef = (element: ReactElement) => {
-  // React <=18 in DEV
   let getter = Object.getOwnPropertyDescriptor(element.props, "ref")?.get;
   let mayWarn = getter && "isReactWarning" in getter && getter.isReactWarning;
   if (mayWarn) {
     return (element as any).ref;
   }
 
-  // React 19 in DEV
   getter = Object.getOwnPropertyDescriptor(element, "ref")?.get;
   mayWarn = getter && "isReactWarning" in getter && getter.isReactWarning;
   if (mayWarn) {
     return element.props.ref;
   }
 
-  // Not DEV
   return element.props.ref || (element as any).ref;
 };
