+++ packages/shared/lib/composeEventHandlers/composeEventHandlers.ts
@@ -15,4 +15,4 @@ const composeEventHandlers = <E>(
   };
 };
 
-export default composeEventHandlers;
+export {composeEventHandlers};
