+++ packages/shared/lib/dispatchDiscreteCustomEvent/index.ts
@@ -1,1 +1,1 @@
-export {default as dispatchDiscreteCustomEvent} from "./dispatchDiscreteCustomEvent";
+export * from "./dispatchDiscreteCustomEvent";
