+++ packages/shared/lib/index.ts
@@ -1,9 +1,9 @@
 export * from "./composeEventHandlers";
 export * from "./composeRefs";
 export * from "./createContext";
+export * from "./createContext";
 export * from "./createContextScope";
 export * from "./dispatchDiscreteCustomEvent";
-export * from "./createContext";
 export * from "./getOpenState";
 export * from "./makeId";
 export * from "./noop";
