+++ packages/shared/lib/tab/getTabbableCandidates.ts
@@ -13,8 +13,9 @@ const getTabbableCandidates = (container: HTMLElement) => {
   const walker = document.createTreeWalker(container, NodeFilter.SHOW_ELEMENT, {
     acceptNode: (node: any) => {
       const isHiddenInput = node.tagName === "INPUT" && node.type === "hidden";
-      if (node.disabled || node.hidden || isHiddenInput)
+      if (node.disabled || node.hidden || isHiddenInput) {
         return NodeFilter.FILTER_SKIP;
+      }
       // `.tabIndex` is not the same as the `tabindex` attribute. It works on the
       // runtime's understanding of tabbability, so this automatically accounts
       // for any kind of element that could be tabbed to.
@@ -23,7 +24,9 @@ const getTabbableCandidates = (container: HTMLElement) => {
         : NodeFilter.FILTER_SKIP;
     },
   });
-  while (walker.nextNode()) nodes.push(walker.currentNode as HTMLElement);
+  while (walker.nextNode()) {
+    nodes.push(walker.currentNode as HTMLElement);
+  }
   // we do not take into account the order of nodes with positive `tabIndex` as it
   // hinders accessibility to have tab order different from visual order.
   return nodes;
