+++ packages/slot/lib/isSlottable.ts
@@ -1,8 +1,9 @@
 import {isValidElement, ReactElement, ReactNode} from "react";
+
 import {Slottable} from "./Slottable";
 
 const isSlottable = (child: ReactNode): child is ReactElement => {
   return isValidElement(child) && child.type === Slottable;
 };
 
-export default isSlottable;
+export {isSlottable};
